Make the dashboard plugin template actually build
